+++ src/ModularPipelines.Build/Modules/LocalMachine/AddLocalNugetSourceModule.cs
@@ -1,10 +1,11 @@
 using ModularPipelines.Attributes;
 using ModularPipelines.Context;
+using ModularPipelines.DotNet.Extensions;
+using ModularPipelines.DotNet.Options;
 using ModularPipelines.Exceptions;
+using ModularPipelines.Extensions;
 using ModularPipelines.Models;
 using ModularPipelines.Modules;
-using ModularPipelines.NuGet.Extensions;
-using ModularPipelines.NuGet.Options;
 
 namespace ModularPipelines.Build.Modules.LocalMachine;
 
@@ -23,7 +24,10 @@ public class AddLocalNugetSourceModule : Module<CommandResult>
     {
         var localNugetPathResult = await GetModule<CreateLocalNugetFolderModule>();
 
-        return await context.NuGet()
-            .AddSource(new NuGetSourceOptions(new Uri(localNugetPathResult.Value!), "ModularPipelinesLocalNuGet"));
+        return await context.DotNet().Nuget.Add.Source(new DotNetNugetAddSourceOptions
+        {
+            PackageSourcePath = localNugetPathResult.Value.AssertExists(),
+            Name = "ModularPipelinesLocalNuGet",
+        });
     }
 }