+++ src/ModularPipelines.Build/Modules/UploadPackagesToNugetModule.cs
@@ -1,15 +1,16 @@
+using EnumerableAsyncProcessor.Extensions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using ModularPipelines.Attributes;
 using ModularPipelines.Build.Settings;
 using ModularPipelines.Context;
+using ModularPipelines.DotNet.Extensions;
+using ModularPipelines.DotNet.Options;
 using ModularPipelines.Extensions;
 using ModularPipelines.Git.Attributes;
 using ModularPipelines.GitHub.Attributes;
 using ModularPipelines.Models;
 using ModularPipelines.Modules;
-using ModularPipelines.NuGet.Extensions;
-using ModularPipelines.NuGet.Options;
 
 namespace ModularPipelines.Build.Modules;
 
@@ -56,10 +57,12 @@ public class UploadPackagesToNugetModule : Module<CommandResult[]>
 
         var packagePaths = await GetModule<PackagePathsParserModule>();
 
-        return await context.NuGet()
-            .UploadPackages(new NuGetUploadOptions(packagePaths.Value!.AsPaths(), new Uri("https://api.nuget.org/v3/index.json"))
+        return await packagePaths.Value!
+            .SelectAsync(async nugetFile => await context.DotNet().Nuget.Push(new DotNetNugetPushOptions
             {
+                Path = nugetFile,
                 ApiKey = _nugetSettings.Value.ApiKey!,
-            });
+            }, cancellationToken), cancellationToken: cancellationToken)
+            .ProcessOneAtATime();
     }
 }