+++ src/ModularPipelines.DotNet/Options/DotNetRestoreOptions.cs
@@ -8,7 +8,7 @@ namespace ModularPipelines.DotNet.Options;
 public record DotNetRestoreOptions : DotNetOptions
 {
     [PositionalArgument(PlaceholderName = "[<ROOT>]")]
-    public string? Root { get; set; }
+    public string? Path { get; set; }
 
     [CommandSwitch("--configfile")]
     public string? Configfile { get; set; }